refactor: Modify api endpoints
